# pta-booking/doc/openapi.yaml
---
openapi: 3.0.3
info:
  title: Booking API
  description: Booking API
  termsOfService: Your terms here
  contact:
    name: Example API Support
    url: https://ptagency.ch/contact
    email: techsupport@ptagency.ch
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
paths:
  /sessions:
    get:
      tags:
      - Session Resource
      parameters:
      - name: from
        in: query
        schema:
          $ref: "#/components/schemas/LocalDate"
      - name: to
        in: query
        schema:
          $ref: "#/components/schemas/LocalDate"
      responses:
        "200":
          description: OK
          content:
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionDto"
  /sessions/{id}:
    get:
      tags:
      - Session Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/SessionId"
      responses:
        "200":
          description: OK
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/SessionDto"
    put:
      tags:
      - Session Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/SessionId"
      responses:
        "200":
          description: OK
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/SessionId"
    post:
      tags:
      - Session Resource
      responses:
        "200":
          description: OK
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/SessionId"
    delete:
      tags:
      - Session Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/SessionId"
      responses:
        "204":
          description: No Content
  /sessions/{id}/bookings:
    get:
      tags:
      - Session Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/SessionId"
      responses:
        "200":
          description: OK
          content:
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingDto"
  /sessions/{id}/bookings/{participant}:
    post:
      tags:
      - Session Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/SessionId"
      - name: participant
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/ParticipantId"
      responses:
        "200":
          description: OK
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/BookingId"
    delete:
      tags:
      - Session Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/SessionId"
      - name: participant
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/ParticipantId"
      responses:
        "204":
          description: No Content
components:
  schemas:
    BookingDto:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/UUID"
        participant:
          $ref: "#/components/schemas/UUID"
    BookingId:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/SessionId"
        participant:
          $ref: "#/components/schemas/ParticipantId"
    LocalDate:
      format: date
      type: string
      example: 2022-03-10
    ParticipantId:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
    SessionDto:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
    SessionId:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string